move Makie support to an extension
